Move storeOf to kstore
