+++ kstore-file/src/commonMain/kotlin/io/github/xxfast/kstore/file/KStore.kt
@@ -1,8 +1,8 @@
 package io.github.xxfast.kstore.file
 
-import io.github.xxfast.kstore.Codec
 import io.github.xxfast.kstore.DefaultJson
 import io.github.xxfast.kstore.KStore
+import io.github.xxfast.kstore.storeOf
 import kotlinx.io.files.Path
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
@@ -27,23 +27,3 @@ public inline fun <reified T : @Serializable Any> storeOf(
   default = default,
   enableCache = enableCache,
 )
-
-/**
- * Creates a store with a given codec
- *
- * @param default returns this value if the file is not found. defaults to null
- * @param enableCache maintain a cache. If set to false, it always reads from disk
- * @param codec codec to be used.
- *
- * @return store that contains a value of type [T]
- */
-public inline fun <reified T : @Serializable Any> storeOf(
-  codec: Codec<T>,
-  default: T? = null,
-  enableCache: Boolean = true,
-): KStore<T> = KStore(
-  default = default,
-  enableCache = enableCache,
-  codec = codec
-)
-
