+++ kstore/src/commonTest/kotlin/io/github/xxfast/kstore/KStoreTests.kt
@@ -12,7 +12,7 @@ import kotlin.test.assertEquals
 import kotlin.test.assertSame
 
 class KStoreTests {
-  private val store: KStore<Cat> = KStore(codec = TestCodec())
+  private val store: KStore<Cat> = storeOf(codec = TestCodec())
 
   @AfterTest
   fun cleanup() {
